Create profile screen
0. Incorporate modifier into state visualization
1. Pass puuid and region from summoner find
2. Implement state holder tagging
3. Create main screen
4. Create summoner view screen with basic switching (using state subholding)
5. Make profile overview a subview of the summoner view (using state subholding)
6. Fill profile overview with extra data like it had in the legacy app
7. Adjust picture url getting from Data Dragon for use with Coil
8. Implement various utilities etc
9. Libraries upgrade

+++ findSummonerImpl/src/main/java/com/tsarsprocket/reportmid/findSummonerImpl/domain/SummonerData.kt
@@ -12,5 +12,5 @@ data class SummonerData(
     val region: Region,
     val riotId: String,
     val iconUrl: String,
-    val level: Long,
+    val level: Int,
 ) : Parcelable
